REF: split policy module into subpackage
