+++ simplesat/sat/policy/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+from .policy import DefaultPolicy
+from .undetermined_clause_policy import (
+    LoggedUndeterminedClausePolicy, UndeterminedClausePolicy
+)
+
+InstalledFirstPolicy = LoggedUndeterminedClausePolicy
